fix 2 parser bugs: vtrans; token prefix in merge
vtrans: update rule name `transform` to vtrans since `transform` in
Transformer is a special function name, which does not merge correctly
in `merge_transformers()` call.

token prefix: there are many places in the parser where the token type
is used. However, the token type is modified by `merge_transformers()`
call with a prefix, so `parse_ecgpattern()` will not work, which uses
`merge_transformers()`.